svg/painting/parsing/paint-order-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing paint-order with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#PaintOrderProperty" rel="help"/><h:meta content="paint-order supports only the grammar 'normal | [ fill || stroke || markers ]'." name="assert"/><h:meta content="paint-order uses the shortest serialization." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("paint-order", "normal stroke");
test_invalid_value("paint-order", "fill fill");
test_invalid_value("paint-order", "markers normal");

  ]]></script></svg>
